feat, refactor: add ensure_use_active_conversation_with_callback everywhere where needed, adjust view_context (ctx) to have active msg
- Move start_handler from handlers/main.py to handlers/start.py
- Make Decimal types instead of float where they must be
- Add username to msgs when order is a gift

+++ bot/keyboards/order.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from bot.callbacks import (
     FixedQuantityCallback, CustomQuantityCallback, BackCallback, BackDestination,
@@ -48,7 +50,7 @@ def build_user_not_found_kb() -> InlineKeyboardMarkup:
         [InlineKeyboardButton("◀️ Назад", callback_data=BackCallback(BackDestination.CHOOSE_RECIPIENT))]
     ])
 
-def build_payment_methods_kb(sbp_price: float, card_price: float, back_dest: BackDestination) -> InlineKeyboardMarkup:
+def build_payment_methods_kb(sbp_price: Decimal, card_price: Decimal, back_dest: BackDestination) -> InlineKeyboardMarkup:
     return InlineKeyboardMarkup([
         [InlineKeyboardButton(f"📲 СБП — {sbp_price} ₽", callback_data=PaymentMethodCallback("sbp"))],
         [InlineKeyboardButton(f"💳 Картой — {card_price} ₽", callback_data=PaymentMethodCallback("card"))],
